% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autotrain.R
\name{autotrain}
\alias{autotrain}
\title{autotrain}
\usage{
autotrain(train, valid, y, x, algorithms = c("deeplearning", "randomForest",
  "gbm"), eval_metric = "AUTO", validation_type = "SharedHoldout",
  runtime_secs = 10, wd = getwd())
}
\arguments{
\item{train}{H2O frame object containing labeled data for model training.
No Default.}

\item{valid}{H2O frame object containing labeled data for model validation.
No Default.}

\item{y}{Character object of length 1 identifying the column name of the target variable. No Default.}

\item{x}{Character object of length 1 or more identifying the column name(s) of the input variables. No Default.}

\item{algorithms}{Character object of length 3, 2, or 1, specifying which alrogrithms to automatically train. The autotrain function will run a separate grid search for each algorimth type. Choices are: "deeplearning", "randomForest", and "gbm" following the naming convention in H2O version 3. Defaults to c("deeplearning", "randomForest", "gbm").}

\item{eval_metric}{Character object defining evaluation metric for training. Defualt is "AUTO" and uses built-in H2O automatic choice for target data type.}

\item{validation_type}{Defines validation type for training models.  Defaults to "shared_holdout" indicating all model built with all algorithms share the same validation set. Currently, this is the only option in autotrain. Planned types include "random_holdout" where each model will get a unique randomized sample of labeled data for validation, and "xval" in which the cross validation functionality in H2O will be implemented in every model.}

\item{runtime_secs}{Character Object which sets the length of time each grid search will run. Defaults to 20, thus the default runtime is 20 sec * (length of algorimths) = 1 minute.}

\item{wd}{Character object defining file path where resulting modeling will be saved. Defualts to current working directory.}
}
\value{
List object containing H2O model objects
}
\description{
Autotrain implements H2O grid search to automatically build machine learning
models
}
