% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl-autogrid.R
\name{dl_autogrid}
\alias{dl_autogrid}
\title{dl_autogrid}
\usage{
dl_autogrid(train, valid, y, x, eval_metric = "AUTO", wd = getwd(),
  folds = NULL, deeplearning_runtime_secs = 10,
  deeplearning_stopping_rounds = 10,
  deeplearning_stopping_tolerance = 1e-05,
  deeplearning_adaptive_rate = TRUE, grid_strategy = "RandomDiscrete")
}
\arguments{
\item{train}{H2O frame object containing labeled data for model training.
No Default.}

\item{valid}{H2O frame object containing labeled data for model validation.
No Default.}

\item{y}{Character object of length 1 identifying the column name of the target variable. No Default.}

\item{x}{Character object of length 1 or more identifying the column name(s) of the input variables. No Default.}

\item{eval_metric}{Character object defining evaluation metric for training. Defualt is "AUTO" and uses built-in H2O automatic choice for target data type.}

\item{wd}{Character object defining file path where dl_models folder will be created and deep learning models saved. Defaults to current working directory.}

\item{folds}{Character object defining number of folds for xval. Default is NULL and currently is not implemented.}

\item{deeplearning_runtime_secs}{Numeric object defining total number of seconds the hyper parameter grid search will run.}

\item{deeplearning_stopping_rounds}{Numeric object defining maximum number of training rounds an individual deep learning model not improving will continue to run. Default is 10.}

\item{deeplearning_stopping_tolerance}{Numeric object which sets the mimmum loss funciton improvement for a training iteration to be considered an improvement. Defulat is 1E-5.}

\item{deeplearning_adaptive_rate}{Boolean, if TRUE ADELTA is used to control learning rate if FALSE than normal rate controls can be used.}

\item{grid_strategy}{Character object default and only current supported option is "randomDiscrete"}
}
\value{
List object containing H2O model objects. Additionally saves h2o models as re-loadable text files in wd/dl_models folder.
}
\description{
dl_autogrid is a wrapper employing built-in settings to run grid search hyper parameter optimizations on deep learning (dl_ algorithms)
}
