% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/even-lengths.R
\name{even_lengths}
\alias{even_lengths}
\title{even_lengths}
\usage{
even_lengths(train_rmse)
}
\arguments{
\item{train_rmse}{List object contiaing model performance numbers. No Default.}
}
\description{
Small, internal function needed for plotting. Not intended for direct use.
}
