% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-frame.R
\name{get_frame}
\alias{get_frame}
\title{get_frame}
\usage{
get_frame(mlout, new_data_id = id, output_models, csv = TRUE, wd)
}
\arguments{
\item{mlout}{startml mlblob object from startml function. No Default.}

\item{new_data_id}{Character object of length 1 identifying the name of id varialbe. No default.}

\item{output_models}{Character object of varying length identifying which model names to get prections from. No default.}

\item{csv}{Boolean, whether or not a csv file named "new_data_predictions_}

\item{wd}{Character object defining file path where dl_models folder will be created and deep learning models saved. Defaults to current working directory.}
}
\value{
R data.frame object with an id column and a single column for each model's predcitions.
}
\description{
get_frame Returns an r data frame having an id and prediction column from a model
}
