% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-hist.R
\name{get_hist}
\alias{get_hist}
\title{get_hist}
\usage{
get_hist(x)
}
\arguments{
\item{x}{model from mlblob. No Default.}
}
\value{
Validation history of the model defined by input
}
\description{
Internal function used in plotting mlblob objects. Retries scoring history from validation given h2o model
}
