% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-ids.R
\name{get_ids}
\alias{get_ids}
\title{get_ids}
\usage{
get_ids(x)
}
\arguments{
\item{x}{model from mlblob. No Default.}
}
\value{
Validation history of the model defined by input
}
\description{
Internal function used in plotting mlblob objects. Gets list of model ids from mlblob object.
}
