% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-model-id.R
\name{get_model_id}
\alias{get_model_id}
\title{get_model_id}
\usage{
get_model_id(x)
}
\arguments{
\item{x}{h2o model. No Default.}
}
\value{
Character object Model id.
}
\description{
Internal function used in plotting mlblob objects. Gets model id from h2o model object
}
