% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-prediction.R
\name{get_prediction}
\alias{get_prediction}
\title{get_prediction}
\usage{
get_prediction(model, mlblob)
}
\arguments{
\item{model}{Character. Name of model including in mlblob object whos predictions will be extracted.}

\item{mlblob}{mlblob object from output of startml function. Defines the mlblob object from which the predictions will be extract. Must have preictions in newdata slot.}
}
\value{
h2oFrame of predictions from model in named in input sourced from the prediction_newdata slot of the named mlblob object.
}
\description{
Retrieves newdata predictions from an exisiting mlblob object from a given model name
}
