% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist-text.R
\name{hist_text}
\alias{hist_text}
\title{hist_text}
\usage{
hist_text(id_final, hist_melted)
}
\arguments{
\item{id_final}{Character Model Id.}

\item{hist_melted}{data.frame object contianing long form training history of H2O model.}
}
\value{
Character object Model id.
}
\description{
Internal function used in plotting mlblob objects. Gets model id from h2o model object
}
