% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-models.R
\name{load_models}
\alias{load_models}
\title{load_models}
\usage{
load_models(path)
}
\arguments{
\item{path}{Character object of file path to folder containing one or more saved h2o model files.}
}
\value{
List object of h2o model objects
}
\description{
Loads saved text file h2o models into R workspace and h2o instance
}
