% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste-nas.R
\name{paste_nas}
\alias{paste_nas}
\title{paste-nas}
\usage{
paste_nas(x, longest)
}
\arguments{
\item{x}{Character or Numeric object No Defalut.}

\item{longest}{Length of character or numeric object to extend x to with nas. No Default.}
}
\value{
Validation history of the model defined by input
}
\description{
Internal function used in plotting mlblob objects. Handles nas.
}
