% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\title{plot}
\usage{
plot(mlout)
}
\arguments{
\item{mlout}{mlblob object from output of startml function.}
}
\value{
None. Plots graphic to device.
}
\description{
Masked from graphics. Summary plot of mlblob object.
}
