% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-blob.R
\name{predict_blob}
\alias{predict_blob}
\title{predict_blob}
\usage{
predict_blob(test, selected_models)
}
\arguments{
\item{test}{H2O frame object used as new data input.}

\item{selected_models}{List of H2O models to predict on new data.}
}
\value{
List of predictions same as length selected_models
}
\description{
Uses all selected models stored in mlblob object to make predictions
}
