% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove-model.R
\name{remove_model}
\alias{remove_model}
\title{remove}
\usage{
remove_model(mlout, model_id)
}
\arguments{
\item{mlout}{mlblob object from startml}

\item{model_id}{Character. Model id of unwanted model as seen in the result of calling plot on the mlblob object.}
}
\value{
mlblob object identical to input except for removal of unwanted model and its predictions.
}
\description{
Removes a specific model from an mlblob object. Useful when trim function leaves an unwanted model in the mlblob.
}
