% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top-models.R
\name{top_models}
\alias{top_models}
\title{top_models}
\usage{
top_models(sorted_models, model_list, number_top_models = 10)
}
\arguments{
\item{sorted_models}{Data frame object indexing all models sorted by decreasing performance. No Default.}

\item{model_list}{List object of H2O frames containing h2o model objects.
No Default.}

\item{number_top_models}{Numeric object indicating number of top models to return. Defualt is 10. If number entered is greater than number of model, whole model list is returned.}
}
\value{
List object same length as number_top_models containing top performing h2o models from model_list.
}
\description{
Selects best n models by performance threshold
}
