% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{trim}
\alias{trim}
\title{trim}
\usage{
trim(mlout, eval_metric, eval_threshold = NULL,
  correlation_threshold = NULL, number_top_models = NULL,
  return_dataframe = FALSE)
}
\arguments{
\item{mlout}{mlblob object from output of startml function.}

\item{eval_metric}{Character object defining evaluation metric for training. No default, use "logloss," "MSE," "RMSE," "MAE," "AUC," or "mean_per_class_error."}

\item{eval_threshold}{Numeric object defining the performance threshold models must meet to be used in prediction. Is minimum for maximization loss function (i.e., AUC) and maximum for minimization loss functions (logloss, MSE, etc). Default is NULL, returns models without performance consideration.}

\item{correlation_threshold}{Numeric object defining the maximum person correlation allowed in the group of resulting models. If two models show high correlation, the one with surperior performance will be kept and the other dropped. Value ranges from -1 to 1, default is NULL, returning models without correlation considered.}

\item{number_top_models}{Numeric object indicating number of top models to return. Defualt is null. If number entered is greater than number of model, whole model list is returned.}

\item{return_dataframe}{Boolean, if TRUE startml will attempt to return a data.frame of the resulting predictions for each new data row. This will only work if the resulting predictions from new data are small enough to be stored in the R workspace. Though, when working with smaller datasets, such as some competitions, this can be very convient. The same object is stored in the H2O space and can be accessed with the name set as the ouput of startml and manipulated with functions from the h2o R package. Default is FALSE.}
}
\value{
mlblob object of selected models from original input.
Trim serves as an option to select only models which meet certain criteria after the startml function is run. Trim inputs the mlblob object from startml and outputs an mlblob object containing only the slected models.
}
\description{
trim
}
