% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid-metric.R
\name{valid_metric}
\alias{valid_metric}
\title{select_models}
\usage{
valid_metric(model_list, eval_metric)
}
\arguments{
\item{model_list}{List object of H2O frames containing h2o model objects. No Default.}

\item{eval_metric}{Character object one of logloss, MSE, RMSE, MAE, AUC, or mean_per_class_error.}
}
\value{
List object same length as model_list containing performance of each model on validation data with selected metric.
}
\description{
select_models
}
